requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
docker-pycreds==0.4.0
einops==0.4.1
frozenlist==1.3.0
fsspec==2022.5.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna==3.3
importlib-metadata==4.12.0
Markdown==3.3.7
multidict==6.0.2
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.2.2
opencv-python==4.6.0.66
packaging==21.3
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
pytorch-lightning==1.6.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.1
sentry-sdk==1.7.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.0
torchaudio==0.12.0
torchmetrics==0.9.2
torchvision==0.13.0
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.10
wandb==0.12.21
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
